python/pypi/bitcoinlib/requirements.txt
bitcoinlib==0.6.15
python-dotenv==1.0.1
